refactor(server): merge ./constants into /libs/env
